requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
Werkzeug==2.3.7
requests==2.31.0
pyserial==3.5